move hook to shared
