+++ frontend/packages/ux-editor/src/components/config/ExpressionContent/Expression/Expression.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { StudioExpressionProps } from '@studio/components';
 import { StudioExpression } from '@studio/components';
-import { useExpressionTexts } from './useExpressionTexts';
+import { useExpressionTexts } from 'app-shared/hooks/useExpressionTexts';
 
 export type ExpressionProps = Omit<StudioExpressionProps, 'texts'>;
